// src/core/icp-brasil/interfaces/icp-brasil.interfaces.ts
export interface IPKCS7SignatureData {
  signatureData: string; // Base64
  signerCertificate: Buffer;
  timestampData?: string;
  signatureAlgorithm: string;
  hashAlgorithm: string;
  signedAttributes: any;
  unsignedAttributes?: any;
}